+++ programs/src/main/java/software/sava/anchor/programs/drift/anchor/types/SerumV3FulfillmentConfig.java
@@ -30,7 +30,7 @@ public record SerumV3FulfillmentConfig(PublicKey _address,
                                        int marketIndex,
                                        SpotFulfillmentType fulfillmentType,
                                        SpotFulfillmentConfigStatus status,
-                                       int[] padding) implements Borsh {
+                                       byte[] padding) implements Borsh {
 
   public static final int BYTES = 344;
   public static final Filter SIZE_FILTER = Filter.createDataSizeFilter(BYTES);
@@ -155,7 +155,7 @@ public record SerumV3FulfillmentConfig(PublicKey _address,
     i += Borsh.len(fulfillmentType);
     final var status = SpotFulfillmentConfigStatus.read(_data, i);
     i += Borsh.len(status);
-    final var padding = Borsh.readArray(new int[4], _data, i);
+    final var padding = Borsh.readArray(new byte[4], _data, i);
     return new SerumV3FulfillmentConfig(_address,
                                         discriminator,
                                         pubkey,
